initial porting to VS 2005
git-svn-id: https://xdxf.svn.sourceforge.net/svnroot/xdxf/trunk@31 9cfc4783-1b1d-0410-aaba-dcc21bd7f26c

+++ src/dictd_generator.cpp
@@ -147,7 +147,8 @@ bool Generator::on_have_data(const StringList& keys, const std::string& data)
 		return false;
 	}
 
-	char *coord=new_string(std::string("\t")+b64_encode(off)+"\t"+b64_encode(size));
+	char *coord = new_string(std::string("\t") + 
+		b64_encode(off) + "\t" + b64_encode(guint32(size)));
 	coord_list.push_back(coord);
 	for (StringList::const_iterator p = keys.begin(); p != keys.end(); ++p)
 		keys_list.push_back(dictd_key(new_string(*p), coord));
